Move GithubIcon to nav dir
